accept upper or lower case in sql code

+++ DBMS.sh
@@ -5,43 +5,61 @@ echo "--------------------------------------------------------------------------
 while true; do
 	read -p "mysql->" REPLY
     case $REPLY in
-        create\ database\ *)
-        	dir_name=$(echo $REPLY | awk -F" " '{print $3}')
-		if [[ -e $dir_name ]]; then
-			echo " this name is used "
-		else
-		
-			if [[ ! "$dir_name" =~ ^[a-zA-Z_]+[a-zA-Z0-9_]+$ ]]; then
-				echo "the name must consist of { _ or character or number } but can't start with number or have a space in it"
+        [cC][rR][eE][aA][tT][eE]\ [dD][aA][tT][aA][bB][aA][sS][eE]\ *)
+        	nf=$(echo $REPLY | awk -F" " '{print NF}')
+        	if [[ $nf -eq 3 ]]; then
+        		
+        		dir_name=$(echo $REPLY | awk -F" " '{print $3}')
+			if [[ -e $dir_name ]]; then
+				echo " this name is used "
 			else
+		
+				if [[ ! "$dir_name" =~ ^[a-zA-Z_]+[a-zA-Z0-9_]+$ ]]; then
+					echo "the name must consist of { _ or character or number } but can't start with number or have a space in it"
+				else
 				
-				mkdir "$dir_name"
-	                	echo "Database '$dir_name' created."
+					mkdir "$dir_name"
+	                		echo "Database '$dir_name' created."
+				fi
 			fi
-		fi
+        		
+        	else
+        		echo "Syntax Error."
+        	fi
+        	
             	;;
-        show\ databases)
+        [sS][hH][oO][wW]\ [dD][aA][tT][aA][bB][aA][sS][eE][sS])
             	echo "Listing All databases:"
             	echo "---------"
             	ls -d */ | sed 's:/$::'
 		echo "---------"
             	;;
-        use\ *)
-		dir_name=$(echo $REPLY | awk -F" " '{print $2}')
-		if [ -d "$dir_name" ]; then
-			clear
-    			./connect.sh $dir_name
+        [uU][sS][eE]\ *)
+        	nf=$(echo $REPLY | awk -F" " '{print NF}')
+        	if [[ $nf -eq 2 ]]; then
+			dir_name=$(echo $REPLY | awk -F" " '{print $2}')
+			if [ -d "$dir_name" ]; then
+				clear
+	    			./connect.sh $dir_name
+			else
+	    			echo "Database  "$dir_name" Not Found"
+			fi
 		else
-    			echo "Database  "$dir_name" Not Found"
-		fi
+        		echo "Syntax Error."
+        	fi
 	       	
             	;;
-        drop\ database\ *)
-        	dir_name=$(echo $REPLY | awk -F" " '{print $3}')
-            	rm -r "$dir_name"
-            	echo "Database '$dir_name' deleted."
+        [dD][rR][oO][pP]\ [dD][aA][tT][aA][bB][aA][sS][eE]\ *)
+        	nf=$(echo $REPLY | awk -F" " '{print NF}')
+        	if [[ $nf -eq 3 ]]; then
+			dir_name=$(echo $REPLY | awk -F" " '{print $3}')
+		    	rm -r "$dir_name"
+		    	echo "Database '$dir_name' deleted."
+            	else
+        		echo "Syntax Error."
+        	fi
            	;;
-        exit)
+        [eE][xX][iI][tT])
             	echo "Exiting the DBMS."
             	exit 
             	;;
